Update bundled coderay gem to use new 'rake gems' infrastructure
The bundled coderay gem was causing a number of errors on application
startup (and when running rake tasks).

  config.gem: Unpacked gem coderay in vendor/gems has no specification
    file. Run 'rake gems:refresh_specs' to fix this.
  config.gem: Unpacked gem coderay in vendor/gems not in a versioned
    directory. Giving up.

Unfortunately, 'rake gems:refresh_specs' wouldn't work, because coderay
didn't have a "versioned directory".  A versioned directory has a name of
the form:

  vendor/gems/mygem-VERSION

The easiest way to manage gems in Rails 2.2.2 is to use 'config.gem'
statements in environment.rb.  So I added the appropriate line to
environment.rb and ran:

  rake gems:unpack

You can see the status of all declared gems using 'rake gems', and there
are a variety of useful rake tasks for working with these declarations.

I haven't touched anything else in the vendor/ directory, because
everything else works as-is with Rails 2.2.2.